requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.17.4
Pillow==6.2.1
psycopg2-binary==2.8.4
requests==2.22.0
urllib3==1.25.7
